ジェネレータ用の基幹モジュールを実装
