+++ lib/rgrb/exec/irc_bot.rb
@@ -101,7 +101,8 @@ module RGRB
           plugin_options[adapter] = {
             root_path: root_path,
             plugin: plugin_config,
-            config_id: config.id
+            config_id: config.id,
+            logger: logger
           }
 
           logger.warn(
