+++ lib/rgrb/plugin/server_connection_report/mail_generator.rb
@@ -2,7 +2,7 @@
 # frozen_string_literal: true
 
 require 'rgrb/version'
-require 'rgrb/plugin/configurable_generator'
+require 'rgrb/generator_plugin'
 
 require 'stringio'
 require 'mail'
@@ -14,7 +14,7 @@ module RGRB
     module ServerConnectionReport
       # メール生成を司るクラス
       class MailGenerator
-        include ConfigurableGenerator
+        include GeneratorPlugin
 
         # メールテンプレートの読み込みに失敗した際に発生するエラー
         class MailTemplateLoadError < StandardError; end
@@ -53,9 +53,6 @@ module RGRB
 
           @mail_config = {}
           @to = 'root@localhost'
-
-          @logger =
-            Lumberjack::Logger.new($stdout, progname: self.class.to_s)
         end
 
         # 設定データを解釈してプラグインの設定を行う
@@ -73,9 +70,6 @@ module RGRB
             to_config = mail_config['To']
             @to = to_config if to_config
           end
-
-          logger = config_data[:logger]
-          @logger = logger if logger
 
           self
         end
@@ -116,13 +110,13 @@ module RGRB
           begin
             load_mail_template(content)
           rescue => e
-            @logger.error("メールテンプレート #{path} の読み込みに失敗しました")
+            logger.error("メールテンプレート #{path} の読み込みに失敗しました")
 
             # 再送出
             raise e
           end
 
-          @logger.warn("メールテンプレート #{path} の読み込みが完了しました")
+          logger.warn("メールテンプレート #{path} の読み込みが完了しました")
 
           self
         end
